refactor: change the name of productDisplayjs to productShowCase.js
