+++ src/js/productShowCase.js
