+++ src/js/searchBar.js
@@ -1,5 +1,5 @@
 import { query } from "./query.js";
-import { productDisplay, removeProductDisplay } from "./productDisplay.js";
+import { productDisplay, removeProductDisplay } from "./productShowCase.js";
 
 const logo = document.querySelector(".logo");
 const slogan = document.querySelector(".slogan");
